manage pytest-docker on a per-session basis
