--- tests/channels/docker-compose.yml
@@ -1,7 +1,0 @@
-version: "3"
-
-services:
-  redis:
-    image: redis
-    ports:
-      - "6397:6379" # use a non-standard port here
